requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.5
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.14.1
eval-type-backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
filetype==1.2.0
frozenlist==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-pinecone==0.1.3
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.0
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.24.4
openai==1.98.0
orjson==3.10.15
packaging==24.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pypdf==5.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-multipart==0.0.20
pyyaml==6.0.2
rapidfuzz==3.9.7
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
starlette==0.44.0
tabulate==0.9.0
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspect==0.9.0
unstructured==0.11.8
unstructured-client==0.28.1
urllib3==2.2.3
uvicorn==0.33.0
wrapt==1.17.2
yarl==1.15.2
